# openapi.yaml
openapi: 3.0.3
info:
  description: "API for managing agents, teams, and tasks in an AutoGen system."
  title: AutoGen Framework API
  version: 1.0.0
servers:
- description: Production environment
  url: https://api.autogenframework.com/v1
- description: Local development environment
  url: http://localhost:3000
paths:
  /agents:
    get:
      description: Retrieves a list of all agents available in the system.
      operationId: listAgents
      parameters:
      - description: Filter agents by status.
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - active
          - inactive
          type: string
        style: form
      - description: Filter agents by role.
        explode: true
        in: query
        name: role
        required: false
        schema:
          type: string
        style: form
      - description: Maximum number of agents to return.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          minimum: 1
          type: integer
        style: form
      - description: Pagination offset.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Agent'
                type: array
          description: A list of agents
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: List all agents
      tags:
      - Agents
    post:
      description: Adds a new agent to the system.
      operationId: createAgent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
          description: Agent created successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Create a new agent
      tags:
      - Agents
  /agents/{agent_id}:
    delete:
      description: Removes an agent from the system by its ID.
      operationId: deleteAgent
      parameters:
      - description: The ID of the agent
        explode: false
        in: path
        name: agent_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Agent deleted successfully
        "404":
          description: Agent not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Delete an agent
      tags:
      - Agents
    get:
      description: Retrieves detailed information about an agent by its ID.
      operationId: getAgentById
      parameters:
      - description: The ID of the agent
        explode: false
        in: path
        name: agent_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
          description: Agent details
        "404":
          description: Agent not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Get details of a specific agent
      tags:
      - Agents
    patch:
      description: Allows partial updates to an existing agent's information.
      operationId: updateAgent
      parameters:
      - description: The ID of the agent
        explode: false
        in: path
        name: agent_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
          description: Agent updated successfully
        "404":
          description: Agent not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listAgents_default_response'
          description: An error response
      summary: Update an existing agent
      tags:
      - Agents
...
components:
...
  schemas:
    Agent:
      example:
        system_message: system_message
        role: role
        agent_id: 0
        name: name
        model_type: model_type
        created_at: 2000-01-23T04:56:07.000+00:00
        status: active
      properties:
        agent_id:
          description: Unique identifier of the agent
          title: agent_id
          type: integer
        name:
          description: The name of the agent
          title: name
          type: string
        role:
          description: "The agent's primary/global role (e.g., 'developer')"
          title: role
          type: string
        status:
          description: The status of the agent
          enum:
          - active
          - inactive
          title: status
          type: string
        model_type:
          description: The underlying model used by this agent
          title: model_type
          type: string
        system_message:
          description: Additional system message or notes about the agent
          title: system_message
          type: string
        created_at:
          description: Timestamp when the agent was created
          format: date-time
          title: created_at
          type: string
      title: Agent
      type: object
    AgentInput:
      example:
        system_message: system_message
        role: role
        name: name
        model_type: model_type
      properties:
        name:
          description: The name of the agent
          title: name
          type: string
        role:
          description: The role or type of the agent
          title: role
          type: string
        model_type:
          description: The underlying model used by this agent
          title: model_type
          type: string
        system_message:
          description: Additional system message or notes about the agent
          title: system_message
          type: string
      required:
      - model_type
      - name
      - role
      title: AgentInput
      type: object
    AgentUpdate:
      description: Schema for partially updating an agent
      example:
        system_message: system_message
        role: role
        name: name
        model_type: model_type
        status: active
      properties:
        name:
          title: name
          type: string
        role:
          title: role
          type: string
        model_type:
          title: model_type
          type: string
        system_message:
          title: system_message
          type: string
        status:
          enum:
          - active
          - inactive
          title: status
          type: string
      title: AgentUpdate
      type: object
...
    listAgents_default_response:
      example:
        code: code
        details: "{}"
        message: message
      properties:
        code:
          description: "A unique error code (e.g., \"AGENT_NOT_FOUND\")"
          title: code
          type: string
        message:
          description: A human-readable error message
          title: message
          type: string
        details:
          description: Additional context or fields related to the error
          title: details
          type: object
      required:
      - code
      - message
      title: listAgents_default_response
      type: object