# openapi_server/models/agent.py
# coding: utf-8

"""
    AutoGen Framework API

    API for managing agents, teams, and tasks in an AutoGen system.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501
from datetime import datetime
from typing import Optional
from sqlmodel import SQLModel, Field

class Agent(SQLModel, table=True):
    agent_id: Optional[int] = Field(default=None, primary_key=True)
    name: str
    role: str
    status: Optional[str] = Field(default="active", nullable=True)
    model_type: str
    system_message: Optional[str] = None
    created_at: Optional[datetime] = None